requirements.txt
Django>=3.2.4,<3.3
djangorestframework==3.14.0
psycopg2>=2.8.6,<=3.0
drf-spectacular>=0.15.1,<0.16